Forbid setting Dataptr methods for ALTLIST
ALTLIST classes do not have Dataptr and
Dataptr_or_null methods. The implementation is
not quite ideal, as it is cumbersome to extend
to other forbidden methods for other classes,
ALTLIST is hardwired in the code. To improve,
I believe that we would need to represent
forbidden methods via the class hierarchy, i.e.
ALTLIST classes should not have Dataptr methods
at all.