Omit \Sexpr[stage=render] warning in install for top level
Currently, if a package has

    \Sexpr[stage=render,results=rd]{"\\\\examples{foo()}"}

then R CMD INSTALL (and thus R CMD check as well) reports a
warning, because \Sexpr{} is not considered to be a legitimate
top level section:

    Warning: dynextest.Rd:14: Section \Sexpr is unrecognized and will be dropped

However, \Sexpr{} as a top level section is ok, if the dynamic
code is executed at render time.

This patch omits the warning for these cases.